pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "columbus"
version = "0.1.0"
description = "Minimal AI agent framework for HUD OSWorld benchmarking"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["ai", "agent", "osworld", "benchmarking"]

dependencies = [
    # Core framework dependencies
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    # LLM framework dependencies  
    "dspy-ai",
    "litellm",
    # Memory and vector database dependencies (optional)
    "mem0ai",
    "qdrant-client",
    # CUA integration
    "cua-agent[hud]",
    "cua-computer[all]",
    # Additional dependencies
    "httpx>=0.25.0",
    "requests>=2.31.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "websockets>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
columbus = "columbus.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["columbus"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "W", "F", "I"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
